feat: 添加地址相关API和删除地址时的参数校验
在AddressBookController中添加了新增地址、设置默认地址、根据id查询地址、查询默认地址、查询指定用户的全部地址的API注释
修改了更新地址的请求方式为POST
修改了删除地址的请求方式为DELETE，并添加了参数校验
在MessageConstant中添加了删除失败的常量
在UserController中添加了删除Redis中验证码的逻辑

+++ take-out-server/src/main/java/com/plumsnow/controller/user/AddressBookController.java
@@ -29,6 +29,7 @@ public class AddressBookController {
      * 新增
      */
     @PostMapping
+    @ApiOperation("新增地址")
     public Result<AddressBook> save(@RequestBody AddressBook addressBook) {
         addressBook.setUserId(BaseContext.getCurrentId());
         log.info("addressBook:{}", addressBook);
@@ -40,16 +41,17 @@ public class AddressBookController {
      * 设置默认地址
      */
     @PutMapping("default")
+    @ApiOperation("设置默认地址")
     public Result<AddressBook> setDefault(@RequestBody AddressBook addressBook) {
         log.info("addressBook:{}", addressBook);
-        LambdaUpdateWrapper<AddressBook> wrapper = new LambdaUpdateWrapper<>();
-        wrapper.eq(AddressBook::getUserId, BaseContext.getCurrentId());
-        wrapper.set(AddressBook::getIsDefault, 0);
-        //SQL:update address_book set is_default = 0 where user_id = ?
-        addressBookService.update(wrapper);
 
-        addressBook.setIsDefault(1);
-        //SQL:update address_book set is_default = 1 where id = ?
+        if(addressBook.getIsDefault() == 1){
+            addressBook.setIsDefault(0);
+        }else {
+            addressBook.setIsDefault(1);
+        }
+
+        //SQL:update address_book set is_default = 1/0 where id = ?
         addressBookService.updateById(addressBook);
         return Result.success(addressBook);
     }
@@ -58,6 +60,7 @@ public class AddressBookController {
      * 根据id查询地址
      */
     @GetMapping("/{id}")
+    @ApiOperation("根据id查询地址")
     public Result get(@PathVariable Long id) {
         AddressBook addressBook = addressBookService.getById(id);
         if (addressBook != null) {
@@ -71,6 +74,7 @@ public class AddressBookController {
      * 查询默认地址
      */
     @GetMapping("default")
+    @ApiOperation("查询默认地址")
     public Result<AddressBook> getDefault() {
         LambdaQueryWrapper<AddressBook> queryWrapper = new LambdaQueryWrapper<>();
         queryWrapper.eq(AddressBook::getUserId, BaseContext.getCurrentId());
@@ -90,6 +94,7 @@ public class AddressBookController {
      * 查询指定用户的全部地址
      */
     @GetMapping("/list")
+    @ApiOperation("查询指定用户的全部地址")
     public Result<List<AddressBook>> list(AddressBook addressBook) {
         addressBook.setUserId(BaseContext.getCurrentId());
         log.info("addressBook:{}", addressBook);
@@ -102,7 +107,7 @@ public class AddressBookController {
         return Result.success(addressBookService.list(queryWrapper));
     }
 
-    @PutMapping
+    @PostMapping("/update")
     @ApiOperation("修改地址")
     public Result<String> edit(@RequestBody AddressBook addressBook){
 
@@ -113,9 +118,12 @@ public class AddressBookController {
         return Result.success(MessageConstant.UPDATE_SUCCESS);
     }
 
-    @DeleteMapping()
+    @DeleteMapping("/{id}")
     @ApiOperation("删除地址")
-    public Result<String> delete(Long id){
+    public Result<String> delete(@PathVariable Long id){
+        if(id == null){
+            return Result.error(MessageConstant.DELETE_FAILURE);
+        }
 
         addressBookService.removeById(id);
